feat: wip fix render media func. and extract to shared
